requirements.txt
annotated-types==0.5.0
asgiref==3.7.2
attrs==23.1.0
beautifulsoup4==4.12.2
cattrs==23.1.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cloudinary==1.33.0
crispy-bootstrap4==2022.1
cryptography==41.0.1
defusedxml==0.7.1
dj-database-url==0.5.0
dj-rest-auth==2.1.9
Django==3.2.19
django-allauth==0.44.0
django-bootstrap-datepicker-plus==5.0.4
django-bootstrap4==23.2
django-cloudinary-storage==0.3.0
django-cors-headers==4.1.0
django-crispy-forms==1.14.0
django-filter==23.2
django-schema-graph==3.1.0
django-taggit==4.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
idna==3.4
oauthlib==3.2.2
Pillow==9.5.0
psycopg2==2.9.6
pycparser==2.21
pydantic==2.0.3
pydantic_core==2.3.0
PyJWT==2.7.0
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==1.26.16
